🐛 fix: 코드래빗 수정사항 변경
