+++ utils/api/calendarApi.ts
@@ -1,59 +1,21 @@
 import { tokenStore } from "@/utils/store/tokenStore";
+import {
+  CalendarData,
+  DailyCalendarData,
+} from "@/utils/types/calendar";
+
+export type { DayStatus, DayInfo, MonthInfo, CalendarData, MealItem, DailyCalendarData } from "@/utils/types/calendar";
 
 const BASE_URL = process.env.EXPO_PUBLIC_API_URL;
 if (!BASE_URL) throw new Error("EXPO_PUBLIC_API_URL 환경변수가 설정되지 않았습니다.");
 
-export type DayStatus = "ENERGETIC" | "NORMAL";
-
-export type DayInfo = {
-  day_of_month: number;
-  status: DayStatus;
-};
-
-export type MonthInfo = {
-  year: number;
-  month: number;
-  start_day_of_month: number;
-  end_day_of_month: number;
-  start_day_of_week: number;
-  end_day_of_week: number;
-};
-
-export type CalendarData = {
-  day_infos: DayInfo[];
-  month_info: MonthInfo;
-};
-
-export type MealItem = {
-  meal_id: number;
-  recorded_at: string | null;
-  meal_time_slot: string;
-  category: string;
-  scan_type: string;
-  food_name: string;
-  kcal: number;
-  carbohydrate_g: number;
-  protein_g: number;
-  fat_g: number;
-  sugar_g: number;
-  sodium_mg: number;
-  memo: string | null;
-};
-
-export type DailyCalendarData = {
-  date: string;
-  total_kcal: number;
-  total_carbohydrate_g: number;
-  total_protein_g: number;
-  total_fat_g: number;
-  total_sugar_g: number;
-  total_sodium_mg: number;
-  breakfast_meals: MealItem[];
-  lunch_meals: MealItem[];
-  dinner_meals: MealItem[];
-  snack_meals: MealItem[];
-  midnight_snack_meals: MealItem[];
-};
+async function parseJson(res: Response): Promise<{ message?: string; data?: unknown }> {
+  try {
+    return await res.json();
+  } catch {
+    return {};
+  }
+}
 
 export async function fetchDailyCalendar(
   date: string,
@@ -70,7 +32,7 @@ export async function fetchDailyCalendar(
     },
   });
 
-  const json = await res.json();
+  const json = await parseJson(res);
   if (!res.ok) throw new Error(json.message ?? "일별 캘린더 조회에 실패했습니다.");
 
   return json.data as DailyCalendarData;
@@ -98,7 +60,7 @@ export async function fetchCalendar(
     },
   });
 
-  const json = await res.json();
+  const json = await parseJson(res);
   if (!res.ok) throw new Error(json.message ?? "캘린더 조회에 실패했습니다.");
 
   return json.data as CalendarData;
